Refactoring whole project, now any bigger implementation has separate application
