+++ Online-shop/urls.py
@@ -25,10 +25,13 @@ urlpatterns = [
     path("orders/",
          include("orders.urls",
                  namespace="orders")),
-    path("", include("shop.urls",
-                     namespace="news")),
-    path("", include("shop.urls",
+    path("",
+         include("shop.urls",
                      namespace="store")),
-    path("", include("shop.urls",
+    path("",
+         include("news.urls",
+                     namespace="news")),
+    path("",
+         include("shop.urls",
                      namespace="main_site")),
 ]
