+++ news/templates/news/post/detail.html
@@ -1,4 +1,4 @@
-{% extends "shop/base.html" %}
+{% extends "main/base.html" %}
 
 {% block title %} {{ post.title }} {% endblock %}
