+++ news/templates/news/post/list.html
@@ -1,4 +1,4 @@
-{% extends "shop/base.html" %}
+{% extends "main/base.html" %}
 
 {% block title %}GEO News{% endblock %}
