+++ orders/templates/order/create.html
@@ -1,4 +1,4 @@
-{% extends "shop/base.html" %}
+{% extends "main/base.html" %}
 
 {% block title %}
     Kasa
