+++ orders/templates/order/created.html
@@ -1,4 +1,4 @@
-{% extends "shop/base.html" %}
+{% extends "main/base.html" %}
 
 {% block title %}
     Zamówienie nr {{ order.id }}
