--- shop/templates/shop/main_site/main.html
@@ -1,1 +1,0 @@
-{% extends "shop/base.html" %}
