+++ shop/templates/shop/product/detail.html
@@ -1,4 +1,4 @@
-{% extends "shop/base.html" %}
+{% extends "main/base.html" %}
 {% load static %}
 
 {% block title %}
