+++ shop/urls.py
@@ -6,11 +6,6 @@ from . import views
 app_name = "shop"
 
 urlpatterns = [
-    path("news/", views.PostListView.as_view(),
-         name="post_list"),
-    path("news/<int:year>/<int:month>/<int:day>/<slug:post>",
-         views.post_detail,
-         name="post_detail"),
     path("store/",
          views.product_list,
          name="product_list"),
